requirements.txt
python-dotenv==1.0.0
openai==1.73.0
elevenlabs==1.56.0
pillow==10.4.0
moviepy==2.1.2
pydub==0.25.1
requests==2.32.3
tqdm==4.67.1 